[refactor] Fix file path
