+++ tests/src/given-component.cpp
@@ -7,7 +7,7 @@
 #include <ranges>
 #include <tuple>
 
-#include "prop-pp/impl/bdd.hpp"
+#include "prop-pp/bdd.hpp"
 
 #include "fixtures.hpp"
