+++ tests/src/when-component.cpp
@@ -9,7 +9,7 @@
 #include <string_view>
 #include <tuple>
 
-#include "prop-pp/impl/bdd.hpp"
+#include "prop-pp/bdd.hpp"
 
 #include "fixtures.hpp"
